requirements.txt
opencv_python
numpy
fastapi
uvicorn
jinja2
tensorflow
pillow
python-multipart
